How to represent bit-array in union?
Exploring how to specify an array of bits in C that would be useful
inside of a Union. I have the feeling that the bitfield syntax
should support this, but I have yet to sort that out. Nor have I
found a handy exmaple. For the moment, I have resorted to using
getter/setter functions, but that feels wrong.